prevent running as root by default + update check
* Add npm upate check at startup
* Prevent the tool from running as root by default
* Add the `--su` CLI option to override root check

The root check will automatically and silently downgrade
permissions for the tool if it detects that it's been
started as root. If downgrade is not possible, an error
will be thrown. This is to help protect the local system
when running arbitrary tests against modules or git
endpoints. The `--su` CLI option can be set to bypass
that default check.