script: re add feature to run a custom script
This commit adds back to feature that allows CITGM to be given an
optional second argument that specifies a file path or link to an
alternative script to be run for testing.

This feature also supports a new option in the manifest "script",
which will allow individuals to specify custom scripts to be used
whenever a module is run (including with citgm-all)

This patch is coming in fully testing with close to 100% coverage.
Currently sitting around 90% on average

There is currently quite a bit of overlap betwee `lib/npm/test`
and `lib/npm/run`, but I think refactoring that would make sense
to happen later, as there is quite a bit of repeated code between
all Child Process stuff, and this could likely be improved.