npm: add ls to the pipeline
There is the possibility that `npm install` will not create the correct tree.
It is also possible that an incorrect tree could potentially cause a process
to hang and never exit while the tests are running.

This commit introduces a new feature to run `npm-ls` and verify the node_modules
tree before running `npm-test`.

All new code has 100% test coverage