npm: fix install error handling
Prior to the code coverage we were checking codes on the exit on npm
install. I was mistaken about the way npm install handles error handling
and had removed the code handling. This brings it back and includes a
test to make sure we do not regress again.

This should solve the CI problems we are having with PPC.