npm: introduce timeout for npm install
We are currently noticing problems in CI with some npm install's
hanging for quite some time. This can make an entire CI run useless.

This PR introduces a new cli options `-o` or `--timeout` which takes
a single argument, a length (of type Number). The default is set to
10 minutes. If an npm install is taking more than 10 minutes it will
simply fail.