flaky: introduce new schema for checking flakyness
originally flakyness could only be set as a boolean.

Now it can be set in multiple ways

As a boolean: true | false
As a string: version | platform | arch | platform-arch
e.g 'v5' | 'darwin' | 'x86' | 'darwin-x86'
As an array of targets: ['v5', 'v4', 'x86']
As an object for more granular control:

{
  'v4': ['darwin', 'arm'],
  'v5': true,
  'v6': 'linux'
}

Or even as an array of objects:

[{ 'linux': 'v6' }, {'darwin': ['v4', 'v5', 'v6']}]

There is potentially too much granularity here... but it all works
and it is all tested with 100% code coverage so why not.