test: fix coverage
currently coverage is broken due to some deep turtles in NYC

I've been holding off doing a hack like this for a bit, but it may be
a little while until we see the edgecase we run into fixed upstream.

This change breaks out the name of the node / npm bin used in
`npm/test` to allow us to instrument it during testing.