lib: use bl to store npm output
When we currently run npm in a childprocess we have been keeping the
output as a string on the context, this is used later to report the
results of the test run.

Rather than storing the information in plain text we can use a
BufferList to store the original buffers, and only call toString()
when dealing with the final buffer.

This PR also abstracted out making a context into a helper function.

Closes #284

PR-URL: https://github.com/nodejs/citgm/pull/291