match conditions: fix for rhel and sles x64/x86
This PR fixes and issue where dotenv was sometimes returning osRelease
with a parsed key and sometimes without. This meant that pulling the
distro and release version from the object was much harder. I have
swapped dotenv out with a simple regex that only grabs the information
that we need.