resolve: use package.gitHead if available
If the package has been published from its git repository, npm adds a
`gitHead` property to the package.json file.
After this commit, the resolver will always use this if it is available
and then fallback to the git tag. This allows to run `citgm module`
with many modules without adding them to the lookup file and also
removes the need for the "prefix" option in most cases.