lookup: use master branch for ember-cli
ember-cli@3.11.0 has failing tests. This means that CITGM fails across
all platforms for ember-cli. The issue is fixed in the ember-cli master
branch. It should be safe to use the master branch, as ember-cli
releases approximately every 6 weeks (coinciding with ember releases)
and is very conservative about changes that would require a semver bump
(again, as is ember--they share the same version number).