lib: allow modules without a script called `test`
Some modules do not call their test script `test`. For example, a
module using Yarn workspaces may define `workspace:test` as their
testing script. We can specify the script to run in our `lookup.json`
but there was a place in our code that was specifically looking for
a script called `test`.