fix: don't ignore engines with yarn v2+ (#831)
yarn v2 doesn't support the `YARN_IGNORE_ENGINES` environment variable.
We still need to set `YARN_IGNORE_ENGINES` for older versions of yarn
to allow installation with unreleased versions of Node.js. Projects are
able to use yarn policies to set the version of yarn they are using,
so we need to version sniff to determine whether or not to set the
environment variable.