grab-module-data: fix broken test
npm 7 and higher now have an exit code for failed fetches. This
introduces a subtle behavior change in testing as the error cases are
slightly different. This tiny change to `grab-module-data.js` adds a
regex to ensure a consistent code path is followed across all
maintained versions of npm.

TBH there is likely a cleaner way to fix this, but it would involve
a lot more refactoring which could create other unexpected bugs.