lib: log correct version and sha for -c option
When using the -c option to supply a specific sha to test CITGM will
erroneously log the latest version of the module. This commit overrides
the incorrect value with the one that is in the downloaded package.json.
Additionally, when supplying -c CITGM will now output the sha for easier
debugging.